Moved HelpCards.tsx to global components folder /src/components/HelpCards
