+++ src/components/helpCards/index.ts
@@ -1,0 +1,2 @@
+import { HelpCards } from './HelpCards';
+export { HelpCards };
