+++ src/components/helpCards/singleCard/SingleCardContent.tsx
