+++ src/components/helpCards/singleCard/transformToSingleCardProps.ts
